Removed warnings. Round now returns a float.
git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@196 5e24083e-1465-11df-9d2a-2365f6df44e6

+++ Kickle/Kickle/Header Files/Utilities.h
@@ -12,6 +12,6 @@
 #endif
 
 // Rounds a floating point number to an integer
-int round( float num );
+float round( float num );
 
 #endif
